+++ shaders/common.glsl
@@ -443,31 +443,6 @@ vec3 distort(vec3 v) {
 
 
 
-vec3 getShadowPos(vec4 viewPos, float lightDot) {
-	vec4 playerPos = gbufferModelViewInverse * viewPos;
-	vec3 shadowPos = (shadowProjection * (shadowModelView * playerPos)).xyz; // convert to shadow screen space
-	float distortFactor = getDistortFactor(shadowPos);
-	float bias = 0.05
-		+ 0.01 / (lightDot + 0.03)
-		+ distortFactor * distortFactor * 0.5;
-	shadowPos = distort(shadowPos, distortFactor); // apply shadow distortion
-	shadowPos = shadowPos * 0.5 + 0.5;
-	shadowPos.z -= bias * 0.02; // apply shadow bias
-	return shadowPos;
-}
-
-vec3 getLessBiasedShadowPos(vec4 viewPos) {
-	vec4 playerPos = gbufferModelViewInverse * viewPos;
-	vec3 shadowPos = (shadowProjection * (shadowModelView * playerPos)).xyz; // convert to shadow screen space
-	float distortFactor = getDistortFactor(shadowPos);
-	shadowPos = distort(shadowPos, distortFactor); // apply shadow distortion
-	shadowPos = shadowPos * 0.5 + 0.5;
-	shadowPos.z -= 0.005 * distortFactor; // apply shadow bias
-	return shadowPos;
-}
-
-
-
 
 
 vec4 getSkylightPercents() {