+++ shaders/main_files/gbuffers_clouds.glsl
@@ -71,17 +71,30 @@ void main() {
 
 #ifdef VSH
 
+#ifdef ISOMETRIC_RENDERING_ENABLED
+	#include "/lib/isometric.glsl"
+#endif
+
 void main() {
 	texcoord = (gl_TextureMatrix[0] * gl_MultiTexCoord0).xy;
 	
-	gl_Position = ftransform();
+	#ifdef ISOMETRIC_RENDERING_ENABLED
+		vec3 worldPos = endMat(gbufferModelViewInverse * (gl_ModelViewMatrix * gl_Vertex));
+		gl_Position = projectIsometric(worldPos);
+	#else
+		gl_Position = ftransform();
+	#endif
 	
 	#ifdef TAA_ENABLED
-		gl_Position.xy += taaOffset * gl_Position.w;
+		#ifdef ISOMETRIC_RENDERING_ENABLED
+			gl_Position.xy += taaOffset * 0.5;
+		#else
+			gl_Position.xy += taaOffset * gl_Position.w;
+		#endif
 	#endif
 	
 	#ifdef FOG_ENABLED
-		vec4 position = gl_Vertex;//gbufferModelViewInverse * (gl_ModelViewMatrix * gl_Vertex);
+		vec4 position = gl_Vertex;
 		getFogData(position.xyz);
 	#endif